[ refactor ] Move debugging functions to separate file
