+++ src/full/Agda/Compiler/JS/Compiler.hs
@@ -38,7 +38,8 @@ import Agda.TypeChecking.Substitute ( absBody )
 import Agda.TypeChecking.Level ( reallyUnLevelView )
 import Agda.TypeChecking.Monad hiding (Global, Local)
 import Agda.TypeChecking.Monad.Builtin
-import Agda.TypeChecking.Monad.Options ( setCommandLineOptions, reportSLn )
+import Agda.TypeChecking.Monad.Debug ( reportSLn )
+import Agda.TypeChecking.Monad.Options ( setCommandLineOptions )
 import Agda.TypeChecking.Reduce ( instantiateFull, normalise )
 import Agda.TypeChecking.Substitute (TelV(..))
 import Agda.TypeChecking.Telescope
