+++ src/full/Agda/TypeChecking/Monad/Benchmark.hs
@@ -15,6 +15,7 @@ import Prelude hiding (print)
 import Agda.Benchmarking
 
 import Agda.TypeChecking.Monad.Base
+import Agda.TypeChecking.Monad.Debug
 import{-# SOURCE #-} Agda.TypeChecking.Monad.Options
 
 import qualified Agda.Utils.Benchmark as B
