+++ src/full/Agda/TypeChecking/Monad/Constraints.hs
@@ -13,6 +13,7 @@ import qualified Data.Set as Set
 
 import Agda.TypeChecking.Monad.Base
 import Agda.TypeChecking.Monad.Closure
+import Agda.TypeChecking.Monad.Debug
 import Agda.TypeChecking.Monad.Options
 
 import Agda.Utils.Lens
