+++ src/full/Agda/TypeChecking/Monad/Local.hs
@@ -9,6 +9,7 @@ import Agda.Syntax.Internal
 import Agda.TypeChecking.Substitute
 import Agda.TypeChecking.Monad.Base
 import Agda.TypeChecking.Monad.Context
+import Agda.TypeChecking.Monad.Debug
 import Agda.TypeChecking.Monad.Env
 import Agda.TypeChecking.Monad.Options
 import Agda.TypeChecking.Free
