+++ src/full/Agda/TypeChecking/Monad/Options.hs
@@ -17,6 +17,7 @@ import System.FilePath
 import Agda.Syntax.Internal
 import Agda.Syntax.Common
 import Agda.Syntax.Concrete
+import {-# SOURCE #-} Agda.TypeChecking.Monad.Debug
 import {-# SOURCE #-} Agda.TypeChecking.Errors
 import Agda.TypeChecking.Monad.Base
 import Agda.TypeChecking.Monad.State
@@ -384,36 +385,3 @@ whenExactVerbosity k n = whenM $ liftTCM $ hasExactVerbosity k n
 {-# SPECIALIZE verboseS :: MonadTCM tcm => VerboseKey -> Int -> tcm () -> tcm () #-}
 verboseS :: HasOptions m => VerboseKey -> Int -> m () -> m ()
 verboseS k n action = whenM (hasVerbosity k n) action
-
--- | Conditionally print debug string.
-{-# SPECIALIZE reportS :: VerboseKey -> Int -> String -> TCM () #-}
-reportS :: (HasOptions m, MonadDebug m)
-        => VerboseKey -> Int -> String -> m ()
-reportS k n s = verboseS k n $ displayDebugMessage n s
-
--- | Conditionally println debug string.
-{-# SPECIALIZE reportSLn :: VerboseKey -> Int -> String -> TCM () #-}
-reportSLn :: (HasOptions m, MonadDebug m)
-          => VerboseKey -> Int -> String -> m ()
-reportSLn k n s = verboseS k n $
-  displayDebugMessage n (s ++ "\n")
-
--- | Conditionally render debug 'Doc' and print it.
-{-# SPECIALIZE reportSDoc :: VerboseKey -> Int -> TCM Doc -> TCM () #-}
-reportSDoc :: MonadTCM tcm => VerboseKey -> Int -> TCM Doc -> tcm ()
-reportSDoc k n d = liftTCM $ verboseS k n $ do
-  displayDebugMessage n . (++ "\n") . show =<< do
-    d `catchError` \ err ->
-      (\ s -> (sep $ map text
-                 [ "Printing debug message"
-                 , k  ++ ":" ++show n
-                 , "failed due to error:" ]) $$
-              (nest 2 $ text s)) <$> prettyError err
-
--- | Print brackets around debug messages issued by a computation.
-{-# SPECIALIZE verboseBracket :: VerboseKey -> Int -> String -> TCM a -> TCM a #-}
-verboseBracket :: (HasOptions m, MonadDebug m, MonadError err m)
-               => VerboseKey -> Int -> String -> m a -> m a
-verboseBracket k n s m = ifNotM (hasVerbosity k n) m $ {- else -} do
-  displayDebugMessage n $ "{ " ++ s ++ "\n"
-  m `finally` displayDebugMessage n "}\n"
