+++ src/full/Agda/TypeChecking/Monad/Sharing.hs
@@ -10,6 +10,7 @@ import Data.Traversable
 
 import Agda.Syntax.Internal
 import Agda.TypeChecking.Monad.Base
+import Agda.TypeChecking.Monad.Debug
 import Agda.TypeChecking.Monad.Options
 import Agda.Utils.Monad
